[WIP] Separating exporter into pieces. Doesn't run.
